# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderCreate": {
        "additionalProperties": false,
        "properties": {
          "customer_email": {
            "anyOf": [
              {
                "maxLength": 254,
                "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Email"
          },
          "customer_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "customer_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "customer_tax_code": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Tax Code"
          },
          "division": {
            "anyOf": [
              {
                "enum": ["OFFICE_EQUIPMENT", "SECURITY", "GENERAL"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "payment_method": {
            "anyOf": [
              {
                "enum": ["CASH", "BANK_TRANSFER"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Method"
          },
          "payment_status": {
            "default": "UNPAID",
            "enum": ["UNPAID", "PAID", "PAY_LATER"],
            "title": "Payment Status",
            "type": "string"
          },
          "priority": {
            "default": "NORMAL",
            "enum": ["LOW", "NORMAL", "HIGH", "URGENT"],
            "title": "Priority",
            "type": "string"
          },
          "requested_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Date"
          },
          "service_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Address"
          },
          "work_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Description"
          }
        },
        "title": "OrderCreate",
        "type": "object"
      },
...
      "OrderPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "OrderPage",
        "type": "object"
      },
...
      "TaskCreate": {
        "additionalProperties": false,
        "properties": {
          "assignee_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Assignee Ids",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "estimated_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Estimated Hours"
          },
          "order_line_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Line Ids"
          },
          "priority": {
            "anyOf": [
              {
                "enum": ["LOW", "NORMAL", "HIGH", "URGENT"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": ["version", "title", "estimated_hours", "due_at", "assignee_ids"],
        "title": "TaskCreate",
        "type": "object"
      },
...
      "TaskListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": ["items"],
        "title": "TaskListOut",
        "type": "object"
      },
...
      "TaskUpdate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "estimated_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Hours"
          },
          "priority": {
            "anyOf": [
              {
                "enum": ["LOW", "NORMAL", "HIGH", "URGENT"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": ["version"],
        "title": "TaskUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SMYou Pro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "orders_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "DRAFT",
                    "PENDING_DISPATCH",
                    "IN_PROGRESS",
                    "AWAITING_CONFIRMATION",
                    "COMPLETED",
                    "REVISION",
                    "CANCELLED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at_desc",
              "enum": ["created_at_desc", "dispatch"],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Danh sách đơn (tìm theo mã/tên khách/SĐT, lọc theo trạng thái)",
        "tags": ["orders"]
      },
      "post": {
        "operationId": "orders_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tạo đơn nháp",
        "tags": ["orders"]
      }
    },
...
    "/api/v1/orders/{order_id}/tasks": {
      "get": {
        "operationId": "orders_tasks_list",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Danh sách đầu việc của đơn",
        "tags": ["dispatch"]
      },
      "post": {
        "operationId": "orders_tasks_create",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | GUARD_FAILED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tạo đầu việc, giao cho một hoặc nhiều kỹ thuật viên",
        "tags": ["dispatch"]
      }
    },
    "/api/v1/orders/{order_id}/tasks/{task_id}": {
      "get": {
        "operationId": "orders_tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chi tiết 1 đầu việc",
        "tags": ["dispatch"]
      },
      "patch": {
        "operationId": "orders_tasks_update",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | GUARD_FAILED | INVALID_TRANSITION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sửa thông tin đầu việc",
        "tags": ["dispatch"]
      }
    },
    "/api/v1/orders/{order_id}/tasks/{task_id}/assignees": {
      "post": {
        "operationId": "orders_tasks_add_assignee",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAddAssignee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | GUARD_FAILED | INVALID_TRANSITION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thêm kỹ thuật viên vào đầu việc",
        "tags": ["dispatch"]
      }
    },
    "/api/v1/orders/{order_id}/tasks/{task_id}/assignees/{assignment_id}/remove": {
      "post": {
        "operationId": "orders_tasks_remove_assignee",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAssigneeRemove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | GUARD_FAILED | INVALID_TRANSITION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gỡ kỹ thuật viên khỏi đầu việc",
        "tags": ["dispatch"]
      }
    },
    "/api/v1/orders/{order_id}/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "orders_tasks_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCancel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | GUARD_FAILED | INVALID_TRANSITION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Huỷ đầu việc",
        "tags": ["dispatch"]
      }
    },
    "/api/v1/orders/{order_id}/tasks/{task_id}/reopen": {
      "post": {
        "operationId": "orders_tasks_reopen",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskReopen"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "problem+json — NOT_FOUND"
          },
          "409": {
            "description": "problem+json — STALE_VERSION | GUARD_FAILED | INVALID_TRANSITION"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mở lại đầu việc đã xong",
        "tags": ["dispatch"]
      }
    },
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "tasks_board",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "NEEDS_ASSIGNEE",
                    "PENDING_ACCEPTANCE",
                    "ACCEPTED",
                    "IN_PROGRESS",
                    "DONE",
                    "CANCELLED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": ["LOW", "NORMAL", "HIGH", "URGENT"],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "assignee_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee Id"
            }
          },
          {
            "in": "query",
            "name": "due_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due From"
            }
          },
          {
            "in": "query",
            "name": "due_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskBoardOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "problem+json — VALIDATION_ERROR"
          }
        },
        "summary": "Bảng đầu việc toàn công ty (lọc trạng thái/ưu tiên/KTV/hạn)",
        "tags": ["dispatch"]
      }
    },
...
  }
}